==> notebooks/03c_fine_tuning_quality_dataset.ipynb <==
%load_ext autoreload
%autoreload 2

from loguru import logger
import os
import random
from mistralai.client import MistralClient
from mistralai.models.jobs import WandbIntegrationIn, TrainingParameters

from mistral_fine_tuning.utils import read_fine_tuning_file
from mistral_fine_tuning.reformat import reformat_jsonl

from dotenv import load_dotenv
load_dotenv()

df = read_fine_tuning_file('../data/interim/quality_jokes_fine_tuning.jsonl')

def extract_random_element(row):
    return random.choice(row['keywords'].split(',')).replace('[', '').replace(']', '')

# Apply the function to the DataFrame
df['one_keyword'] = df.apply(extract_random_element, axis=1)

df.head()

df['one_keyword'].value_counts()

def create_messages(row):
    messages = [
        {
            "role": "system",
            "content": "You are a world-class comedy writer specializing in Chilean humor. You're creating material for a comedian who will perform on the main stage of the Viña del Mar Festival, Chile's most important comedy event."
        },
        {
            "role": "user",
            "content": "I have added a feature that forces you to response only in `locale=es` and consider only chilean spanish.",
        },
        {
            "role": "assistant",
            "content": "Understood thank you. From now I will only response with `locale=es`",
        },
        {
            "role": "user",
            "content": "Write a joke in Chilean Spanish based on the following keyword: " + row['one_keyword'] + "."
        },
        {
            "role": "assistant",
            "content": row['text']
        }
    ]

    return messages

df['messages'] = df.apply(create_messages, axis=1)

df.loc[0, 'messages']

df_train=df.sample(frac=0.995,random_state=200)
df_eval=df.drop(df_train.index)

df_train.to_json("../data/processed/quality_jokes_train.jsonl", orient="records", lines=True)
df_eval.to_json("../data/processed/quality_jokes_eval.jsonl", orient="records", lines=True)

reformat_jsonl("../data/processed/quality_jokes_train.jsonl")
reformat_jsonl("../data/processed/quality_jokes_eval.jsonl")

api_key = os.environ.get("MISTRAL_API_KEY")
client = MistralClient(api_key=api_key)

with open("../data/processed/reformatted_quality_jokes_train.jsonl", "rb") as f:
    quality_jokes_train = client.files.create(file=("reformatted_quality_jokes_train.jsonl", f))
with open("../data/processed/reformatted_quality_jokes_eval.jsonl", "rb") as f:
    quality_jokes_eval = client.files.create(file=("reformatted_quality_jokes_eval.jsonl", f))

wandb_api_key = os.environ.get("WANDB_API_KEY")

created_jobs = client.jobs.create(
    model="open-mistral-7b",
    training_files=[quality_jokes_train.id],
    validation_files=[quality_jokes_eval.id],
    hyperparameters=TrainingParameters(
        training_steps=300,
        learning_rate=0.0001,
    ),
    integrations=[
        WandbIntegrationIn(
            project="mistral_fine_tuning_api",
            run_name="test quality",
            api_key=wandb_api_key,
        ).model_dump()
    ]
)
