==> notebooks/04_use_fine_tuned_model.ipynb <==
# Using fine-tuned models


import os
import weave
from mistralai.async_client import MistralAsyncClient
from mistralai.models.chat_completion import ChatMessage

from dotenv import load_dotenv
load_dotenv()

client = MistralAsyncClient(api_key=os.environ["MISTRAL_API_KEY"])

@weave.op()
async def call_mistral(model:str, messages:list, temperature:float, **kwargs) -> str:
    "Call the Mistral API"
    chat_response = await client.chat(
        model=model,
        temperature=temperature,
        messages=messages,
        **kwargs,
    )
    return chat_response.choices[0].message.content

@weave.op()
async def generate_joke(keyword:str, model:str, temperature: float) -> str:
    "Generate a new joke"
     
    messages = [ChatMessage(role="system", content="You are a world-class comedy writer specializing in Chilean humor. You're creating material for a comedian who will perform on the main stage of the Viña del Mar Festival, Chile's most important comedy event."),
              ChatMessage(role="user", content="I have added a feature that forces you to response only in `locale=es` and consider only chilean spanish."),
              ChatMessage(role="assistant", content="Understood thank you. From now I will only response with `locale=es`"),
              ChatMessage(role="user", content=f"Write a joke in Chilean Spanish based on the following keyword: '{keyword}'."),
            ]

    joke = await call_mistral(model=model, messages=messages, temperature=temperature)
    return {"keyword": keyword, "joke": joke}

list_models_response = await client.list_models()
for model in list_models_response.data:
    print(model)

weave.init("mistral_hackathon")
out = await generate_joke(keyword="familia", model="open-mistral-7b", temperature=0)
print(out)

out = await generate_joke(keyword="familia", model="ft:open-mistral-7b:fd6d41e7:20240627:eabccc57", temperature=0.5)
print(out)
